+++ extra/topup.php
@@ -14,7 +14,7 @@
 // You should have received a copy of the GNU General Public License
 // along with Moodle.  If not, see <https://www.gnu.org/licenses/>.
 
-/** The topup wallet page Display confiramtion about payment to charge wallet.
+/** The topup wallet page Display confirmation about payment to charge wallet.
  *
  * @package     enrol_wallet
  * @copyright   2023 Mohammad Farouk <phun.for.physics@gmail.com>
@@ -80,7 +80,7 @@ if (confirm_sesskey()) {
         ];
         $message = get_string('confirmpayment', 'enrol_wallet', $a);
 
-        // This code is required form payment button.
+        // This code is required for payment button.
         $code = 'require([\'core_payment/gateways_modal\'], function(modal) {
             modal.init();
         });';